#4 :) - time to celebrate
i will consider this feature done, although a proper reporitng is not in
place the cli reports on the number of images it transfered. As per the
website that owns the dataset it ~17.5k images. The CLI reports 17,459.
The whole dataset is a little under 1GB but performance is already
impressive.
As for the reporting, that will be split out into it's own feature.